requirements.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# AI & ML
openai==1.3.0
requests==2.31.0

# Security & Cryptography
bcrypt==4.0.1
cryptography==41.0.7

# Text Processing
regex==2023.10.3
python-Levenshtein==0.21.1
pyahocorasick==2.0.0

# Utilities
click==8.1.7
Werkzeug==2.3.7
Jinja2==3.1.2

# Development
pytest==7.4.3
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0